refactor: 조건부 렌더링을 상단에서 먼저 처리하는 방식으로 변경
